% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_figure.R
\name{build_figure}
\alias{build_figure}
\title{Build PDF Figure for 5x3 Statistical Analysis Charts}
\usage{
build_figure(param, param2 = NULL, legend = 2, arrangement = F)
}
\arguments{
\item{param}{List containing chart parameters (see Details)}

\item{param2}{Optional second parameter list for two-panel figures (default: NULL)}

\item{legend}{If two-chart figure, use legend from first chart if 1, second chart if 2, or both if 3}

\item{arrangement}{Layout option: T = vertical stack with legend on right,F = horizontal row (default: 2)}
}
\value{
nothing - called for its creation of a plot
}
\description{
Function to compose figures such as those in Aisbett et al.,
"5x3: A Practical Approach to Encouraging Thoughtful Statistical Analysis"
}
\details{
The param and param2 lists should contain:
\describe{
\item{alpha}{Matrix with first row containing alpha values, other rows "*" if one-sided test at that level}
\item{power}{Matrix of powers for sample size estimation, else 0}
\item{sample}{Total sample size (provisional if computing). Invalid values default to 50}
\item{MML, MMU}{Boundaries of minimum meaningful effect magnitudes}
\item{xmin, xmax}{Range of effect sizes to display}
\item{zmin, zmax}{Range of SEs or sample sizes to display}
\item{chartType}{"Design" for sample size estimation stage, otherwise analysis stage}
\item{Study}{Fraction of sample in larger group (two-group study) or 1 (single-group)}
\item{chartBW}{TRUE for black & white chart}
\item{colorvec}{Vector of colors for regions and data points}
\item{ltyBW} line type for B&W charts
\item{ES} number of comma separated list of effect sizes
\item{dataV} variance of data (or possible SE in Analysis stage)
\item{VES} TRUE if SE rather than variance
\item{labels} terminology for the 5 tests
\item{levels} terminology for the test levels (up to 3)
}
}
