% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTables.R
\name{cleanTables}
\alias{cleanTables}
\title{Cllean and reorganise tables of alpha levels and power}
\usage{
cleanTables(alpha, power, MM, level)
}
\arguments{
\item{alpha}{-matrix table in which first row is alphas and  entries are "*" or null. Other entries will be converted to nulls}

\item{power}{-matrix of power(0 if in analysis phase)}

\item{MM}{-difference between upper and lower boundaries of interval equivalent to zero/not materially significant}

\item{level}{\itemize{
\item name for finding at each level
}

Clean up test table.

Unless a column is headed by a number between 0 and .5, blank out the column
Ensure composite test is at the same alpha
delete columns with no tests specified
if MM= 0, delete everything but tests for non-superiority and non-inferiority}
}
\description{
Function to reorder tables so larger column numbers have lower alphas, and all alphas are between 0 and 0.5.
}
