% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDisplay.R
\name{datadisplay}
\alias{datadisplay}
\title{FUNCTION to display data as integers on chart generated in buildPlot.}
\usage{
datadisplay(
  datamean,
  dataV,
  VES,
  sample,
  r,
  ptcol,
  cex,
  xmin,
  xmax,
  ymin,
  ymax
)
}
\arguments{
\item{datamean}{Comma-separated text of study effect sizes.}

\item{dataV}{Comma-separated text of study SEs or variances}

\item{VES}{true if variances in above list}

\item{sample}{total sample size}

\item{r}{\itemize{
\item fraction of sample in larger group if 2 group
}}

\item{ptcol}{color to use for numbered points}

\item{cex}{size of numeral used to plot data}

\item{xmin}{effect size range}

\item{xmax}{ditto}

\item{ymin}{SE range}

\item{ymax}{ditto}
}
\description{
study data are plotted as points

plot points on chart of effect size vs SE.
}
