% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawRegions.R
\name{drawRegionsVT}
\alias{drawRegionsVT}
\title{Draw Rejection Regions for Multi-Level T-Tests}
\usage{
drawRegionsVT(
  chartType,
  alphas,
  power,
  MML,
  MMU,
  xmin,
  xmax,
  ymin,
  ymax,
  ytick,
  DOF,
  colorvec,
  chartBW,
  ltyBW
)
}
\arguments{
\item{chartType}{"Design" for sample size estimation phase, otherwise analysis phase}

\item{alphas}{5 x 3 matrix of alpha levels, with possible NA entries}

\item{power}{5 x 3 matrix of power levels (zero unless in Design phase)}

\item{MML}{Lower boundary of minimum meaningful effect magnitude}

\item{MMU}{Upper boundary of minimum meaningful effect magnitude}

\item{xmin}{Minimum x-axis value to display (effect size)}

\item{xmax}{Maximum x-axis value to display (effect size)}

\item{ymin}{Minimum y-axis value to display (in SE units, even in Design phase)}

\item{ymax}{Maximum y-axis value to display (in SE units)}

\item{ytick}{Vector of y-axis values at which tick marks are placed}

\item{DOF}{Degrees of freedom for t-tests}

\item{colorvec}{Vector of colors for regions (color or B&W version)}

\item{chartBW}{TRUE if black & white chart}

\item{ltyBW}{Vector of line types for non-inferiority/non-superiority in B&W charts}
}
\description{
Depicts rejection regions for various t-tests on a chart.
When computing sample size (Design phase), these regions must achieve
the desired power. The tests depend on how many test levels have been
requested and whether there are effects equivalent to zero.
}
