% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_Fig2a.R
\docType{data}
\name{Fig2a}
\alias{Fig2a}
\title{Fig2a}
\format{
A data frame with  rows and  variables:
\describe{
\item{ chartType }{  character }
\item{ alpha }{  matrix,array }
\item{ power }{  matrix,array }
\item{ sample }{  numeric }
\item{ Study }{  numeric }
\item{ MML }{  numeric }
\item{ MMU }{  numeric }
\item{ xmin }{  numeric }
\item{ xmax }{  numeric }
\item{ zmin }{  numeric }
\item{ zmax }{  numeric }
\item{ levels }{  character }
\item{ labels }{  expression }
\item{ colorvec }{  character }
\item{ ltyBW }{  character }
\item{ ES }{  character }
\item{ dataV }{  character }
\item{ VES }{  logical }
\item{ chartBW }{  logical }
}
}
\source{
Example data, see column 4 of charts.csv
}
\usage{
Fig2a
}
\description{
Parameters to generate Fig 2a in {5x3:A Practical Approach to Encouraging Thoughtful Statistical Analyses}.
}
\keyword{datasets}
