% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_Fig5a.R
\docType{data}
\name{Fig5a}
\alias{Fig5a}
\title{Fig5a}
\format{
A data frame with  rows and  variables:
\describe{
\item{ chartType }{  character }
\item{ alpha }{  matrix,array }
\item{ power }{  matrix,array }
\item{ sample }{  numeric }
\item{ Study }{  numeric }
\item{ MML }{  numeric }
\item{ MMU }{  numeric }
\item{ xmin }{  numeric }
\item{ xmax }{  numeric }
\item{ zmin }{  numeric }
\item{ zmax }{  numeric }
\item{ levels }{  character }
\item{ labels }{  character }
\item{ colorvec }{  character }
\item{ ltyBW }{  character }
\item{ ES }{  character }
\item{ dataV }{  character }
\item{ VES }{  logical }
\item{ chartBW }{  logical }
}
}
\source{
Beanland et al plus example data, see column 8 of charts.csv. Additional documentation available in
the package folder or accessible through the Shiny app’s Help tab.
}
\usage{
Fig5a
}
\description{
Parameters to generate Fig 5a in {5x3:A Practical Approach to Encouraging Thoughtful Statistical Analyses} .
}
\keyword{datasets}
