% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_Fig5b.R
\docType{data}
\name{Fig5b}
\alias{Fig5b}
\title{Fig5b}
\format{
A data frame with  rows and  variables:
\describe{
\item{ chartType }{  character }
\item{ alpha }{  matrix,array }
\item{ power }{  matrix,array }
\item{ sample }{  numeric }
\item{ Study }{  numeric }
\item{ MML }{  numeric }
\item{ MMU }{  numeric }
\item{ xmin }{  numeric }
\item{ xmax }{  numeric }
\item{ zmin }{  numeric }
\item{ zmax }{  numeric }
\item{ levels }{  character }
\item{ labels }{  character }
\item{ colorvec }{  character }
\item{ ltyBW }{  character }
\item{ ES }{  character }
\item{ dataV }{  character }
\item{ VES }{  logical }
\item{ chartBW }{  logical }
}
}
\source{
Beanlane et al plus example data, see column 9 of charts.csv.
}
\usage{
Fig5b
}
\description{
Parameters to generate Fig 5b in {5x3:A Practical Approach to Encouraging Thoughtful Statistical Analyses}
available through the package folder or the Shiny app’s Help tab.
}
\keyword{datasets}
