% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_Fig7.R
\docType{data}
\name{Fig7}
\alias{Fig7}
\title{Fig7}
\format{
A data frame with  rows and  variables:
\describe{
\item{ chartType }{  character }
\item{ alpha }{  matrix,array }
\item{ power }{  matrix,array }
\item{ sample }{  numeric }
\item{ Study }{  numeric }
\item{ MML }{  numeric }
\item{ MMU }{  numeric }
\item{ xmin }{  numeric }
\item{ xmax }{  numeric }
\item{ zmin }{  numeric }
\item{ zmax }{  numeric }
\item{ levels }{  character }
\item{ labels }{  character }
\item{ colorvec }{  character }
\item{ ltyBW }{  character }
\item{ ES }{  character }
\item{ dataV }{  character }
\item{ VES }{  logical }
\item{ chartBW }{  logical }
}
}
\source{
Data from Ladewski & Al-Bayati, see column 11 of charts.csv
}
\usage{
Fig7
}
\description{
Parameters to generate Fig 7 in {5x3:A Practical Approach to Encouraging Thoughtful Statistical Analyses}.
}
\keyword{datasets}
