% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPolygonPoints.R
\name{plotpoints}
\alias{plotpoints}
\title{FUNCTION to ensure corner points used to draw polygons (regions) in function drawregions aren't beyond chart boundaries.}
\usage{
plotpoints(xm, ym, xmin, xmax, ymin, ymax, MM, a, power, DOF)
}
\arguments{
\item{xm}{The x co-ordinate of the corner point of the polygon.}

\item{ym}{The y co-ordinate of the corner point of the polygon.}

\item{xmin, xmax}{range of effect sizes}

\item{ymin, ymax}{range of SE}

\item{MM}{margin from which line starts}

\item{a}{The test level of this polygon.}

\item{power}{Desired power in sample size calculation.}

\item{DOF}{of t-test.}
}
\value{
Co-ordinates of polygon which don't extend beyond chart.
}
\description{
FUNCTION to ensure corner points used to draw polygons (regions) in function drawregions aren't beyond chart boundaries.
}
