% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointData.R
\name{presentDecision}
\alias{presentDecision}
\title{Presents summary data for a point selected on the chart created in buildPlot}
\usage{
presentDecision(
  xm,
  ym,
  alphas,
  chartType,
  power,
  MML,
  MMU,
  DOF,
  levels,
  labels
)
}
\arguments{
\item{xm}{co-ordinates of points (xm = effect size, y = SE (even in sample size computation).}

\item{ym}{as above}

\item{alphas}{6 x 3 matrix with alpha levels in first row, "*" entry where a test at the alpha isdesired.}

\item{chartType}{Design or Analysis}

\item{power}{(zero unless doing sample size calculation).}

\item{MML}{Minimum meaningful effect boundaries}

\item{MMU}{as above}

\item{DOF}{for t-tests.}

\item{levels}{terms for strength of tests at different alphas/findings at different p values}

\item{labels}{terms for tests/findings}
}
\description{
FUNCTION to determine which rejection region a selected point lies in

When computing sample size, these regions must have the desired power.

The tests depend on how many test levels have been requested and whether there are effects equivalent to 0.
}
