% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny app, launching the browser
#initial commit -adding project fies}
\usage{
run_app()
}
\description{
Run the Shiny app, launching the browser
#initial commit -adding project fies
}
