% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPlot.R
\name{buildPlot}
\alias{buildPlot}
\title{Build Chart}
\usage{
buildPlot(param)
}
\arguments{
\item{param}{List containing all chart parameters (see details)}
}
\description{
Creates a complete chart displaying rejection regions for
multi-level statistical tests. Handles both Design phase (sample size
estimation) and Analysis phase visualizations.
}
\details{
The param list should contain:
\describe{
\item{alpha}{Matrix with first row containing alpha values, other rows "*" if one-sided test at that level}
\item{power}{Matrix of powers for sample size estimation, else 0}
\item{sample}{Total sample size (provisional if computing). Invalid values default to 50}
\item{var}{Estimated variance for sample size estimation}
\item{MML}{Boundaries of minimum meaningful effect magnitudes}
\item{MMU}{ditto}
\item{xmin}{Range of effect sizes to display}
\item{xmax}{ditto}
\item{zmin}{Range of SEs or sample sizes to display}
\item{zmax}{ditto}
\item{chartType}{"Design" for sample size estimation phase, otherwise analysis phase}
\item{Study}{Fraction of sample in larger group (two-group study) or 1 (single-group)}
\item{chartBW}{TRUE for black & white chart}
\item{colorvec}{Vector of colors for regions and data points}
\item{ltyBW}{line type for B&W charts}
\item{ES}{number of comma separated list of effect sizes}
\item{dataV}{variance of data in Design phase, data SE or variance in Analysis phase}
\item{VES}{TRUE if variance rather than SE in Analysis data}
}
}
