% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickPoints.R
\name{clickPointsVT}
\alias{clickPointsVT}
\title{Function to allow users to get co-ordinates of a point on click, and in Analysis stage, to compute 95\% CI around that point.}
\usage{
clickPointsVT(xhit, yhit, bi, stage, Study)
}
\arguments{
\item{xhit}{Selected x co-ordinate.}

\item{yhit}{Selected y co-ordinate.}

\item{bi}{\itemize{
\item sample size if analysis stage, else estimated variance .
}}

\item{stage}{"Design" or "Analysis".}

\item{Study}{Type of study (one or two group).}
}
\value{
rounded and possibly converted x-y co-ordinates, and CI boundaries if analysis stage
}
\description{
The calculations for the chart use SE as y-co-ordinate. Translate this to sample size if on sample size tab..
}
