% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawCIs.R
\name{drawCIs}
\alias{drawCIs}
\title{Draw Multi-Level Confidence Intervals}
\usage{
drawCIs(param, lab = "effect size", horizontal = TRUE, legendPos = "topleft")
}
\arguments{
\item{param}{List containing  parameters (see details)}

\item{lab}{Label for the effect size axis (default: "effect size")}

\item{horizontal}{TRUE to orient CIs parallel to x-axis (default: TRUE)}

\item{legendPos}{Legend position using standard plot values (default: "topleft")}
}
\description{
Depicts multilevel CIs for effect size measurements.
CIs corresponding to different test levels are drawn over each other,
with the weakest tests (shortest CIs) shown with the thickest lines.
}
\details{
The param list should contain:
\describe{
\item{alpha}{Matrix with first row containing alpha values, other rows "*" if one-sided test at that level}
\item{sample}{Total sample size (provisional if computing). Invalid values default to 50}
\item{var}{Estimated variance for sample size estimation}
\item{MML}{Boundaries of minimum meaningful effect magnitudes}
\item{MMU}{ditto}
\item{xmin}{Range of effect sizes to display}
\item{xmax}{ditto}
\item{zmin}{Range of SEs or sample sizes to display}
\item{zmax}{ditto}
\item{chartType}{"Design" for sample size estimation phase, otherwise analysis phase}
\item{Study}{Fraction of sample in larger group (two-group study) or 1 (single-group)}
\item{chartBW}{TRUE for black & white chart}
\item{levels}{vector of names for strengths of tests}
\item{ES}{number of comma separated list of effect sizes}
\item{dataV}{variance of data in Design phase, data SE or variance in Analysis phase}
\item{VES}{TRUE if variance rather than SE in Analysis data}
}
}
