% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawLegend.R
\name{drawLegend}
\alias{drawLegend}
\title{Draw Legend for Multi-Level T est Charts}
\usage{
drawLegend(param)
}
\arguments{
\item{param}{List containing  parameters (see details)}
}
\value{
The base plot object for the legend
}
\description{
Creates a composite legend for displaying test results at multiple
significance levels. The legend is composed of multiple sub-legends, one for
each test type, with entries corresponding to test levels.
}
\details{
The param list should contain:
\describe{
\item{alpha}{Matrix with first row containing alpha values, other rows "*" if one-sided test at that level}
\item{power}{Matrix of powers for sample size estimation, else 0}
\item{sample}{Total sample size (provisional if computing). Invalid values default to 50}
\item{var}{Estimated variance for sample size estimation}
\item{MML}{Boundaries of minimum meaningful effect magnitudes}
\item{MMU}{ditto}
\item{xmin}{Range of effect sizes to display}
\item{xmax}{ditto}
\item{zmin}{Range of SEs or sample sizes to display}
\item{zmax}{ditto}
\item{chartType}{"Design" for sample size estimation stage, otherwise analysis stage}
\item{Study}{Fraction of sample in larger group (two-group study) or 1 (single-group)}
\item{chartBW}{TRUE for black & white chart}
\item{colorvec}{Vector of colors for regions and data points}
\item{ltyBW}{line type for B&W charts}
}
}
