% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_params_Excel.R
\name{read_params_Excel}
\alias{read_params_Excel}
\title{read parameters from Excel}
\usage{
read_params_Excel(filename)
}
\arguments{
\item{filename}{name of .csv}
}
\description{
read parameters from Excel
}
